useState-hook-react
The useState hook is a fundamental feature in React, a popular JavaScript library for building user interfaces. It allows developers to manage and manipulate state within functional components, which are a key part of React's component-based architecture.

In React, state represents data that can change over time and affect how a component renders. Prior to hooks, state management was primarily handled in class components. However, with the introduction of hooks in React 16.8, functional components gained the ability to manage their own state.

The useState hook is straightforward to use. It is imported from the 'react' library and called within a functional component. When invoked, it returns an array with two elements: the current state value and a function to update that state. The first element is typically assigned to a variable, while the second is often given a name like setState.